Module for Wave Generating functions
