+++ PSLab_Original/functions.c
@@ -716,220 +716,6 @@ void setFlashPointer(BYTE location){
 
 }
 
-void sqr1(unsigned int wavelength,unsigned int high_time,BYTE scaling) {
-    /*-----------------------square wave output-----------------*/
-    OC1R = high_time-1;
-    PR1 = wavelength-1;
-
-    OC1CON1 = 6; //Edge aligned PWM
-    OC1CON1bits.OCTSEL = 4;
-    T1CONbits.TCKPS = scaling&0x3;
-    OC1CON2 = 11; //01011 = Timer1 synchronizes or triggers OCx (default)
-    T1CONbits.TON = 1;
-
-    if((scaling&0x4)==0)RPOR5bits.RP54R = 0x10; //square wave pin(RC6) mapped to 0b010001 (output compare 1 )
-    /*-----------------------square wave output-----------------*/
-
-}
-
-void sineWave1(unsigned int wavelength,BYTE highres){
-    /*-----------------------sine wave output-----------------*/
-    T4CONbits.TON= 0;
-    _DMA2IF = 0;    _DMA2IE = 0;
-    DMA2CONbits.CHEN = 0;
-    if(highres&1){
-        OC3R =HIGH_RES_WAVE>>1;
-        OC3RS = HIGH_RES_WAVE;
-    }else{
-        OC3R =LOW_RES_WAVE>>1;
-        OC3RS = LOW_RES_WAVE;        
-    }
-    OC3CON2 = 0;
-    OC3CON1 = 6; //Edge aligned PWM
-    OC3CON2bits.SYNCSEL = 0x1F;  //OCRS compare used for sync
-    OC3CON1bits.OCTSEL = 7;     //Fp used as clock
-
-    
-    DMA2CONbits.AMODE = 0; // Configure DMA for Register Indirect mode with post-increment
-    DMA2CONbits.SIZE = 0;  //word transfer
-    DMA2CONbits.MODE = 0; // Configure DMA for Continuous mode
-    DMA2CONbits.DIR = 1; // RAM-to-Peripheral data transfers
-    DMA2PAD = (volatile unsigned int)&OC3R; // Point DMA to OC3R
-    DMA2REQ = 0b11011;    //timer 4 triggers DMA
-
-
-    if(highres&1){
-        DMA2STAH = __builtin_dmapage (&sineTable1);
-        DMA2STAL = __builtin_dmaoffset (&sineTable1);
-        DMA2CNT = WAVE_TABLE_FULL_LENGTH-1; // total table size -1/  DMA requests
-    }else{
-        DMA2STAH = __builtin_dmapage (&sineTable1_short);
-        DMA2STAL = __builtin_dmaoffset (&sineTable1_short);
-        DMA2CNT = WAVE_TABLE_SHORT_LENGTH-1; // total table size -1/  DMA requests
-    }
-
-
-    _DMA2IF = 0;    _DMA2IE = 1;
-    DMA2CONbits.CHEN = 1;
-
-    T4CONbits.TCKPS = (highres>>1)&3;
-    PR4 = wavelength;
-    T4CONbits.TON= 1;
-    
-    RPOR6bits.RP57R = 0x12;  //Sine 1 Mapping output to square wave 4 SQR4
-}
-void sineWave2(unsigned int wavelength,BYTE highres){
-    /*-----------------------sine wave output-----------------*/
-    T3CONbits.TON= 0; T2CONbits.T32=0;
-    _DMA3IF = 0;    _DMA3IE = 0;
-    DMA3CONbits.CHEN = 0;
-    
-    if(highres&1){
-        OC4R =HIGH_RES_WAVE>>1;
-        OC4RS = HIGH_RES_WAVE;
-    }else{
-        OC4R =LOW_RES_WAVE>>1;
-        OC4RS = LOW_RES_WAVE;        
-    }
-
-    OC4CON2 = 0;
-    OC4CON1 = 6; //Edge aligned PWM
-    OC4CON2bits.SYNCSEL = 0x1F;  //OCRS compare used for sync
-    OC4CON1bits.OCTSEL = 7;     //Fp used as clock
-
-    
-    DMA3CONbits.AMODE = 0; // Configure DMA for Register Indirect mode with post-increment
-    DMA3CONbits.SIZE = 0;  //word transfer
-    DMA3CONbits.MODE = 0; // Configure DMA for Continuous mode
-    DMA3CONbits.DIR = 1; // RAM-to-Peripheral data transfers
-    DMA3PAD = (volatile unsigned int)&OC4R; // Point DMA to OC4R
-    DMA3REQ = 0b1000;    //timer 3 triggers DMA
-
-    if(highres&1){
-        DMA3CNT = WAVE_TABLE_FULL_LENGTH-1; // total table size -1/  DMA requests
-        DMA3STAH = __builtin_dmapage (&sineTable2);
-        DMA3STAL = __builtin_dmaoffset (&sineTable2);
-    }else{
-        DMA3CNT = WAVE_TABLE_SHORT_LENGTH-1; // total table size -1/  DMA requests
-        DMA3STAH = __builtin_dmapage (&sineTable2_short);
-        DMA3STAL = __builtin_dmaoffset (&sineTable2_short);
-    }
-
-    _DMA3IF = 0;    _DMA3IE = 1;
-    DMA3CONbits.CHEN = 1;
-
-    T3CONbits.TCKPS = (highres>>1)&3;
-    PR3 = wavelength;
-    T3CONbits.TON= 1;
-    
-    RPOR6bits.RP56R = 0x13;  //Sine 2 Mapping output to square wave 3 SQR3
-}
-
-void setSineWaves(unsigned int wavelength1,unsigned int wavelength2,unsigned int pos,unsigned int tmr_delay,BYTE highres){
-    /*-----------------------sine wave output-----------------*/
-    T3CONbits.TON= 0;T2CONbits.T32= 0; T4CONbits.TON= 0;
-    _DMA3IF = 0;    _DMA3IE = 0;_DMA2IF = 0;    _DMA2IE = 0;
-    DMA2CONbits.CHEN = 0;DMA3CONbits.CHEN = 0;
-    
-    if(highres&1){        //highres = (prescaler1<<4)|(prescaler2<<2)|(HIGHRES2<<1)|(HIGHRES)
-        OC3R =HIGH_RES_WAVE>>1;
-        OC3RS = HIGH_RES_WAVE;
-    }else{
-        OC3R =LOW_RES_WAVE>>1;
-        OC3RS = LOW_RES_WAVE;        
-    }
-
-    
-    if(highres&2){
-        OC4R =HIGH_RES_WAVE>>1;
-        OC4RS = HIGH_RES_WAVE;
-    }else{
-        OC4R =LOW_RES_WAVE>>1;
-        OC4RS = LOW_RES_WAVE;        
-    }
-    
-    
-    
-    OC3CON2 = 0;
-    OC3CON1 = 6; //Edge aligned PWM
-    OC4CON2 = 0;
-    OC4CON1 = 6; //Edge aligned PWM
-    OC3CON2bits.SYNCSEL = 0x1F;  //OCRS compare used for sync
-    OC3CON1bits.OCTSEL = 7;     //Fp used as clock
-    OC4CON2bits.SYNCSEL = 0x1F;  //OCRS compare used for sync
-    OC4CON1bits.OCTSEL = 7;     //Fp used as clock
-
-    
-
-    DMA2CONbits.AMODE = 0; // Configure DMA for Register Indirect mode with post-increment
-    DMA2CONbits.SIZE = 0;  //word transfer
-    DMA2CONbits.MODE = 0; // Configure DMA for Continuous mode
-    DMA2CONbits.DIR = 1; // RAM-to-Peripheral data transfers
-    DMA2PAD = (volatile unsigned int)&OC3R; // Point DMA2 to OC3R
-    DMA2REQ = 0b11011;    //Timer 4 requests DMA
-    
-    
-    DMA3CONbits.AMODE = 0; // Configure DMA for Register Indirect mode with post-increment
-    DMA3CONbits.SIZE = 0;  //word transfer
-    DMA3CONbits.MODE = 0; // Configure DMA for Continuous mode
-    DMA3CONbits.DIR = 1; // RAM-to-Peripheral data transfers
-    DMA3PAD = (volatile unsigned int)&OC4R; // Point DMA3 to OC4R
-    DMA3REQ = 0b1000;    //Timer 3 requests DMA
-
-    if(highres&1){
-        DMA2CNT = WAVE_TABLE_FULL_LENGTH-1; // total table size -1/  DMA requests
-        DMA2STAH = __builtin_dmapage (&sineTable1);
-        DMA2STAL = __builtin_dmaoffset (&sineTable1);
-    }else{
-        DMA2CNT = WAVE_TABLE_SHORT_LENGTH-1; // total table size -1 [  DMA requests]
-        DMA2STAH = __builtin_dmapage (&sineTable1_short);
-        DMA2STAL = __builtin_dmaoffset (&sineTable1_short);
-    }
-
-    
-    if(highres&2){
-        DMA3CNT = WAVE_TABLE_FULL_LENGTH-1; // total table size -1/  DMA requests
-        DMA3STAH = __builtin_dmapage (&sineTable2);
-        DMA3STAL = __builtin_dmaoffset (&sineTable2);
-    }else{
-        DMA3CNT = WAVE_TABLE_SHORT_LENGTH-1; // total table size -1 [  DMA requests]
-        DMA3STAH = __builtin_dmapage (&sineTable2_short);
-        DMA3STAL = __builtin_dmaoffset (&sineTable2_short);
-    }
-    
-
-    _DMA2IF = 0;    _DMA2IE = 1;_DMA3IF = 0;    _DMA3IE = 1;
-    DMA2CONbits.CHEN = 1;DMA3CONbits.CHEN = 1;
-
-    for(i=0;i<pos;i++){DMA3REQbits.FORCE=1;while(DMA3REQbits.FORCE);} //skip these many points for wave #2
-    T3CONbits.TCKPS = (highres>>2)&3;T4CONbits.TCKPS = (highres>>4)&3;
-    PR4 = wavelength1;PR3 = wavelength2;
-    TMR4 = 0;
-    TMR3=tmr_delay;
-
-    RPOR6bits.RP56R = 0x13;  //Sine 2 Mapping output to square wave 3 SQR3
-    RPOR6bits.RP57R = 0x12;  //Sine 1 Mapping output to square wave 4 SQR4
-
-    T3CONbits.TON= 1;T4CONbits.TON= 1;
-    
-}
-
-void sqr2(unsigned int wavelength,unsigned int high_time,BYTE scaling){
-    /*-----------------------square wave output-----------------*/
-    OC4R = high_time-1;
-    PR3 = wavelength-1;
-    OC4CON1 = 6; //Edge aligned PWM
-    OC4CON1bits.OCTSEL = 1;
-    T4CONbits.TCKPS = scaling;
-    OC4CON2 = 13; //01101 = Timer3 synchronizes or triggers OCx (default)
-    T4CONbits.TON= 1;
-
-    RPOR5bits.RP55R = 0x13;  //square wave pin(RC7) mapped to 0x010011 (output compare 4 )
-    /*-----------------------square wave output-----------------*/
-
-}
-
-
 void mapReferenceOscillator(BYTE channel,BYTE scaler){
     /*--------------External clock monitoring, RP55/RC7  ----------*/
     REFOCONbits.ROON = 0;
@@ -945,67 +731,6 @@ void mapReferenceOscillator(BYTE channel,BYTE scaler){
         RPOR3bits.RP41R = 0x31; 
     }
     REFOCONbits.ROON = 1;   
-}
-
-void sqr4(uint16 w,uint16 R0,uint16 R1,uint16 RS1,uint16 R2,uint16 RS2,uint16 R3,uint16 RS3,BYTE scaling){
-    /*-----------------------square wave output-----------------*/
-    _DMA2IF = 0;    _DMA2IE = 0; _DMA3IF = 0;    _DMA3IE = 0;
-    DMA2CONbits.CHEN = 0;DMA3CONbits.CHEN = 0;
-
-    T1CONbits.TON = 0;
-    if((scaling>>6)&1){
-        EnableComparator();
-    OC1CON2 = 0x19; OC1CON2bits.OCTRIG = 1; //COMP2 triggers
-    OC2CON2 = 0x19; OC2CON2bits.OCTRIG = 1;
-    OC3CON2 = 0x19; OC3CON2bits.OCTRIG = 1;
-    OC4CON2 = 0x19; OC4CON2bits.OCTRIG = 1;
-
-    }else{
-    OC1CON2 = 11; //01011 = Timer1 synchronizes or triggers OCx (default)
-    OC2CON2 = 11; //01011 = Timer1 synchronizes or triggers OCx (default)
-    OC3CON2 = 11; //01011 = Timer1 synchronizes or triggers OCx (default)
-    OC4CON2 = 11; //01011 = Timer1 synchronizes or triggers OCx (default)
-    }
-    PR1 = w;
-    OC1R = 0;
-    OC1RS = R0;
-    OC2R = R1;
-    OC2RS = RS1;
-    OC3R = R2;
-    OC3RS = RS2;
-    OC4R = R3;
-    OC4RS = RS3;
-
-    if((scaling>>5)&1){
-        OC1CON1 = 7;  OC2CON1 = 7;   OC3CON1 = 7;   OC4CON1 = 7; //Continuous center aligned PWM. high when R, low when RS.
-        OC1CON2bits.OCINV = 0;
-        OC2CON2bits.OCINV = 0;
-        OC3CON2bits.OCINV = 0;
-        OC4CON2bits.OCINV = 0;
-    }
-    else{
-        OC1CON1 = 4;  OC2CON1 = 4;   OC3CON1 = 4;   OC4CON1 = 4; //double compare mode. one shot toggle on alternate matches of OCR, OCRS
-        OC1CON2bits.OCINV = 0;
-        OC2CON2bits.OCINV = (scaling>>2)&1;
-        OC3CON2bits.OCINV = (scaling>>3)&1;
-        OC4CON2bits.OCINV = (scaling>>4)&1;
-    }
-    OC1CON1bits.OCTSEL = 4;
-    OC2CON1bits.OCTSEL = 4;
-    OC3CON1bits.OCTSEL = 4;
-    OC4CON1bits.OCTSEL = 4;
-
-    T1CONbits.TCKPS = scaling&0x3;
-
-    TMR1=0;
-    T1CONbits.TON = 1;
-
-    RPOR5bits.RP54R = 0x10; //SQR1(RC6) mapped to (output compare 1 )
-    RPOR5bits.RP55R = 0x11; //SQR2(RC7) mapped to (output compare 2 )
-    RPOR6bits.RP56R = 0x12; //SQR3(RC8) mapped to (output compare 3 )
-    RPOR6bits.RP57R = 0x13; //SQR4(RC9) mapped to (output compare 4 )
-    /*-----------------------square wave output-----------------*/
-
 }
 
 void preciseDelay(int t){
