+++ PSLab_Original/nbproject/Makefile-default.mk
@@ -57,17 +57,17 @@ OBJECTDIR=build/${CND_CONF}/${IMAGE_TYPE}
 DISTDIR=dist/${CND_CONF}/${IMAGE_TYPE}
 
 # Source Files Quoted if spaced
-SOURCEFILES_QUOTED_IF_SPACED=proto2_main.c PSLAB_UART.c PSLAB_I2C.c Common_Functions.c PSLAB_NRF.c PSLAB_SPI.c PSLAB_ADC.c
+SOURCEFILES_QUOTED_IF_SPACED=proto2_main.c PSLAB_UART.c PSLAB_I2C.c Common_Functions.c PSLAB_NRF.c PSLAB_SPI.c PSLAB_ADC.c Wave_Generator.c
 
 # Object Files Quoted if spaced
-OBJECTFILES_QUOTED_IF_SPACED=${OBJECTDIR}/proto2_main.o ${OBJECTDIR}/PSLAB_UART.o ${OBJECTDIR}/PSLAB_I2C.o ${OBJECTDIR}/Common_Functions.o ${OBJECTDIR}/PSLAB_NRF.o ${OBJECTDIR}/PSLAB_SPI.o ${OBJECTDIR}/PSLAB_ADC.o
-POSSIBLE_DEPFILES=${OBJECTDIR}/proto2_main.o.d ${OBJECTDIR}/PSLAB_UART.o.d ${OBJECTDIR}/PSLAB_I2C.o.d ${OBJECTDIR}/Common_Functions.o.d ${OBJECTDIR}/PSLAB_NRF.o.d ${OBJECTDIR}/PSLAB_SPI.o.d ${OBJECTDIR}/PSLAB_ADC.o.d
+OBJECTFILES_QUOTED_IF_SPACED=${OBJECTDIR}/proto2_main.o ${OBJECTDIR}/PSLAB_UART.o ${OBJECTDIR}/PSLAB_I2C.o ${OBJECTDIR}/Common_Functions.o ${OBJECTDIR}/PSLAB_NRF.o ${OBJECTDIR}/PSLAB_SPI.o ${OBJECTDIR}/PSLAB_ADC.o ${OBJECTDIR}/Wave_Generator.o
+POSSIBLE_DEPFILES=${OBJECTDIR}/proto2_main.o.d ${OBJECTDIR}/PSLAB_UART.o.d ${OBJECTDIR}/PSLAB_I2C.o.d ${OBJECTDIR}/Common_Functions.o.d ${OBJECTDIR}/PSLAB_NRF.o.d ${OBJECTDIR}/PSLAB_SPI.o.d ${OBJECTDIR}/PSLAB_ADC.o.d ${OBJECTDIR}/Wave_Generator.o.d
 
 # Object Files
-OBJECTFILES=${OBJECTDIR}/proto2_main.o ${OBJECTDIR}/PSLAB_UART.o ${OBJECTDIR}/PSLAB_I2C.o ${OBJECTDIR}/Common_Functions.o ${OBJECTDIR}/PSLAB_NRF.o ${OBJECTDIR}/PSLAB_SPI.o ${OBJECTDIR}/PSLAB_ADC.o
+OBJECTFILES=${OBJECTDIR}/proto2_main.o ${OBJECTDIR}/PSLAB_UART.o ${OBJECTDIR}/PSLAB_I2C.o ${OBJECTDIR}/Common_Functions.o ${OBJECTDIR}/PSLAB_NRF.o ${OBJECTDIR}/PSLAB_SPI.o ${OBJECTDIR}/PSLAB_ADC.o ${OBJECTDIR}/Wave_Generator.o
 
 # Source Files
-SOURCEFILES=proto2_main.c PSLAB_UART.c PSLAB_I2C.c Common_Functions.c PSLAB_NRF.c PSLAB_SPI.c PSLAB_ADC.c
+SOURCEFILES=proto2_main.c PSLAB_UART.c PSLAB_I2C.c Common_Functions.c PSLAB_NRF.c PSLAB_SPI.c PSLAB_ADC.c Wave_Generator.c
 
 
 CFLAGS=
@@ -143,6 +143,13 @@ ${OBJECTDIR}/PSLAB_ADC.o: PSLAB_ADC.c  nbproject/Makefile-${CND_CONF}.mk
 	${MP_CC} $(MP_EXTRA_CC_PRE)  PSLAB_ADC.c  -o ${OBJECTDIR}/PSLAB_ADC.o  -c -mcpu=$(MP_PROCESSOR_OPTION)  -MMD -MF "${OBJECTDIR}/PSLAB_ADC.o.d"      -g -D__DEBUG -D__MPLAB_DEBUGGER_PK3=1  -mno-eds-warn  -omf=elf -DXPRJ_default=$(CND_CONF)  -no-legacy-libc  $(COMPARISON_BUILD)  -O1 -msmart-io=1 -Wall -msfr-warn=off  
 	@${FIXDEPS} "${OBJECTDIR}/PSLAB_ADC.o.d" $(SILENT)  -rsi ${MP_CC_DIR}../ 
 	
+${OBJECTDIR}/Wave_Generator.o: Wave_Generator.c  nbproject/Makefile-${CND_CONF}.mk
+	@${MKDIR} "${OBJECTDIR}" 
+	@${RM} ${OBJECTDIR}/Wave_Generator.o.d 
+	@${RM} ${OBJECTDIR}/Wave_Generator.o 
+	${MP_CC} $(MP_EXTRA_CC_PRE)  Wave_Generator.c  -o ${OBJECTDIR}/Wave_Generator.o  -c -mcpu=$(MP_PROCESSOR_OPTION)  -MMD -MF "${OBJECTDIR}/Wave_Generator.o.d"      -g -D__DEBUG -D__MPLAB_DEBUGGER_PK3=1  -mno-eds-warn  -omf=elf -DXPRJ_default=$(CND_CONF)  -no-legacy-libc  $(COMPARISON_BUILD)  -O1 -msmart-io=1 -Wall -msfr-warn=off  
+	@${FIXDEPS} "${OBJECTDIR}/Wave_Generator.o.d" $(SILENT)  -rsi ${MP_CC_DIR}../ 
+	
 else
 ${OBJECTDIR}/proto2_main.o: proto2_main.c  nbproject/Makefile-${CND_CONF}.mk
 	@${MKDIR} "${OBJECTDIR}" 
@@ -192,6 +199,13 @@ ${OBJECTDIR}/PSLAB_ADC.o: PSLAB_ADC.c  nbproject/Makefile-${CND_CONF}.mk
 	@${RM} ${OBJECTDIR}/PSLAB_ADC.o 
 	${MP_CC} $(MP_EXTRA_CC_PRE)  PSLAB_ADC.c  -o ${OBJECTDIR}/PSLAB_ADC.o  -c -mcpu=$(MP_PROCESSOR_OPTION)  -MMD -MF "${OBJECTDIR}/PSLAB_ADC.o.d"      -mno-eds-warn  -g -omf=elf -DXPRJ_default=$(CND_CONF)  -no-legacy-libc  $(COMPARISON_BUILD)  -O1 -msmart-io=1 -Wall -msfr-warn=off  
 	@${FIXDEPS} "${OBJECTDIR}/PSLAB_ADC.o.d" $(SILENT)  -rsi ${MP_CC_DIR}../ 
+	
+${OBJECTDIR}/Wave_Generator.o: Wave_Generator.c  nbproject/Makefile-${CND_CONF}.mk
+	@${MKDIR} "${OBJECTDIR}" 
+	@${RM} ${OBJECTDIR}/Wave_Generator.o.d 
+	@${RM} ${OBJECTDIR}/Wave_Generator.o 
+	${MP_CC} $(MP_EXTRA_CC_PRE)  Wave_Generator.c  -o ${OBJECTDIR}/Wave_Generator.o  -c -mcpu=$(MP_PROCESSOR_OPTION)  -MMD -MF "${OBJECTDIR}/Wave_Generator.o.d"      -mno-eds-warn  -g -omf=elf -DXPRJ_default=$(CND_CONF)  -no-legacy-libc  $(COMPARISON_BUILD)  -O1 -msmart-io=1 -Wall -msfr-warn=off  
+	@${FIXDEPS} "${OBJECTDIR}/Wave_Generator.o.d" $(SILENT)  -rsi ${MP_CC_DIR}../ 
 	
 endif
