+++ PSLab_Original/nbproject/configurations.xml
@@ -12,6 +12,7 @@
       <itemPath>PSLAB_NRF.h</itemPath>
       <itemPath>PSLAB_SPI.h</itemPath>
       <itemPath>PSLAB_ADC.h</itemPath>
+      <itemPath>Wave_Generator.h</itemPath>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
                    displayName="Linker Files"
@@ -27,6 +28,7 @@
       <itemPath>PSLAB_NRF.c</itemPath>
       <itemPath>PSLAB_SPI.c</itemPath>
       <itemPath>PSLAB_ADC.c</itemPath>
+      <itemPath>Wave_Generator.c</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
