+++ PSLab_Original/proto2_main.c
@@ -22,6 +22,7 @@
 #include "PSLAB_NRF.h"
 #include "PSLAB_SPI.h"
 #include "PSLAB_ADC.h"
+#include "Wave_Generator.h"
 //_FICD(ICS_PGD2 & JTAGEN_OFF) //Programmming pins  ..PGED2
